% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{calcRoot_F}
\alias{calcRoot_F}
\title{Claculate the Roots of F}
\usage{
calcRoot_F(pars)
}
\arguments{
\item{pars}{named vector of parameters found in \code{\link{ecoStep}}. At a minimum, qE needs to be supplied. Optionally, other parameter arguments can be supplied here, but if they aren't, they are taken from the formals (defaults) of \code{\link{ecoStep}}.}
}
\value{
numeric value(s) corresponding to roots of F
}
\description{
Calculate the roots of F (planktivorous fish) by solving a polynomial
}
\examples{
calcRoot_F(pars=c(qE=0.5))
}

