% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{dF_dt_1state}
\alias{dF_dt_1state}
\title{Temporal dynamics of F}
\usage{
dF_dt_1state(F0, pars)
}
\arguments{
\item{pars}{named vector of parameters found in \code{\link{ecoStep}}. At a minimum, qE needs to be supplied. Optionally, other parameter arguments can be supplied here, but if they aren't, they are taken from the formals (defaults) of \code{\link{ecoStep}}.}

\item{F}{abundance of planktivorous fish}
}
\value{
numeric value indicating temporal rate of change of F
}
\description{
Temporal dynamics of F, but as a function of F and parameters (not other state variables)
}
\examples{
getRoot(c(A0=1, F0=1, J0=1), pars=c(qE=0.5)) # find equilibria when bass are rare
dF_dt_1state(F0=17.890184, pars=c(qE=0.5)) # F0 set to equilibrium when bass start as rare

getRoot(c(A0=1000, F0=1, J0=1000), pars=c(qE=0.5)) # find equilibria when bass are abundant
dF_dt_1state(F0=0.06712064, pars=c(qE=0.5)) # F0 set to equilibrium when bass start as abundant

getRoot(c(A0=1, F0=1, J0=1), pars=c(qE=0.65)) # find equilibria when bass are rare
dF_dt_1state(F0=100, pars=c(qE=0.65)) # F0 set to equilibrium when bass start as rare # WTF?!
fishStep(X=c(A0=6.275129e-18, F0=100, J0=1.443280e-17))

getRoot(c(A0=1000, F0=1, J0=1000), pars=c(qE=0.65)) # find equilibria when bass are abundant
dF_dt_1state(F0=0.09136212, pars=c(qE=0.65)) # check planktivore equation for rate of change
fishStep(X=c(A0=364.51517642, F0=0.09136212, J0=838.38490576)) # re-examine rates of change of fish near equilibrium
}

