% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{dU_dF}
\alias{dU_dF}
\title{Change in F temporal dynamics wrt F}
\usage{
dU_dF(F, pars)
}
\arguments{
\item{F}{abundance of planktivorous fish}

\item{pars}{named vector of parameters found in \code{\link{ecoStep}}. At a minimum, qE needs to be supplied. Optionally, other parameter arguments can be supplied here, but if they aren't, they are taken from the formals (defaults) of \code{\link{ecoStep}}.}
}
\value{
numeric value that is du/dt
}
\description{
Define a function u=dF/dt, this gives du/dF. The function u is a function of F and parameters.
}
\examples{
dU_dF(3.702216, pars=c(qE=0.5))
}

