% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoStep.R
\name{ecoStep}
\alias{ecoStep}
\title{Ecosystem Step}
\usage{
ecoStep(X, pars = c(qE = 0.01), fA = 2, cJA = 0.1, cJF = 0.5,
  cFA = 0.3, vuln = 80, hide = 80, surv = 0.6, Fo = 200, DF = 0.09,
  Zmix = 4, nZ = 10, rP = 3, Load = 0.6, mP = 0.1, Ho = 1,
  DH = 0.5, cHF = 0.1, alf = 0.3, cPH = 0.25)
}
\arguments{
\item{X}{vector of length 3 with names P0 (initial phytoplankton), H0 (initial herbivore [zooplankton]), A0 (adult piscivorous fish), F0 (planktivorous fish), and J0 (juvenile piscivores).  Indicates starting value of state variables}

\item{pars}{a length 1 vector with name qE, indicating harvesting rate of adults}

\item{fA}{Fecundity of adult piscivore (2 in OLD)}

\item{cJA}{Density dependent mortality rate of juveniles}

\item{cJF}{Consumption of juveniles by planktivores}

\item{cFA}{Consumption of planktivores by adult piscivores}

\item{vuln}{Vulnerability coefficient (this is "v" in eco lett table/ equations)}

\item{hide}{Hiding coefficient (this is "h" in eco lett table/ equations)}

\item{surv}{Overwinter survivorship of adults}

\item{Fo}{Refuge density of planktivores}

\item{DF}{Diffusion parameter for planktivores}

\item{Zmix}{depth of mixed layer in meters}

\item{nZ}{number of steps for vertical integration}

\item{rP}{phytoplankton growth parameter}

\item{Load}{daily phosphorus load}

\item{mP}{phytoplankton daily mortality}

\item{Ho}{refuge biomass}

\item{DH}{diffusion parameter for zooplankton (herbivore)}

\item{cHF}{consumption rate by planktivore (of zooplankton)}

\item{alf}{conversion efficiency of consumed phytoplankton to zooplankton}

\item{cPH}{consumption rate of phytoplankton by zooplankton}

\item{I0}{surface irradiance, microEinsteins m-1 s-1}

\item{k_sat, k_inh}{photosynthesis-irradiance parameters}

\item{DOC, eps0, epsP}{Light extinction parameters from Carpenter et al. L&O 1998; DOC is dissolved organic carbon; eps are in units per m}
}
\value{
named vector of length 3, with names corresponding to juvenile bass (J0), adult bass (A0), and sunfish (F0) abundances.
}
\description{
Compute the rate of change of the five ecosystem state variables given initial conditions and parameters
}
\note{
\code{F} are referred to as planktivorous fish, but they do feed on the juvenile piscivores (\code{J}).
}
\examples{
ecoStep(X=c(A0=1.5,F0=25,J0=2), pars=c(qE=0.1))
}

