% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{findCritHarv}
\alias{findCritHarv}
\title{Find Critical Values}
\usage{
findCritHarv(parRange = c(0.15, 1.5), stateRange = c(0, 2000),
  par_nGrid = 300, state_nGrid = 5e+05)
}
\arguments{
\item{parRange}{range of parameter (qE) values over which to search}

\item{stateRange}{range of state variable (J0) values over which to search}

\item{par_nGrid}{size of grid of parameter (qE) values}

\item{state_nGrid}{size of grid of state variable (J0) values}
}
\value{
a vector of critical values of qE
}
\description{
Find the critical values of harvest in the trophic cascade model
}
\examples{
 findCritHarv(par_nGrid=1E2, state_nGrid=1E3) # fast but very inaccurate
findCritHarv(par_nGrid=1E2, state_nGrid=5E4) # closer, but still off
}

