% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishStep.R
\name{fishStep}
\alias{fishStep}
\title{Fish Step}
\usage{
fishStep(X, pars)
}
\arguments{
\item{X}{named vector of length 3, or a 1-by-3 matrix, indicating starting value of state variables}

\item{pars}{--- there will be arguments for input parameters}
}
\value{
matrix of 1 row and 3 columns, with names corresponding to juvenile bass, adult bass, and sunfish abundances.
}
\description{
Compute the rate of change of the three fish state variables given initial conditions and parameters
}

