% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishStep.R
\name{fs2D}
\alias{fs2D}
\title{Wrapper for model when finding stability properties}
\usage{
fs2D(t, y, parameters, ...)
}
\arguments{
\item{t}{time, for compatibility with \code{phaseR} and \code{deSolve} and \code{rootSolve}}

\item{y}{state variables as in \code{\link{fishStep2D}}}

\item{parameters}{same as \code{pars} in \code{\link{fishStep2D}}; had to be called 'parameters' for compatibility with \code{phaseR}}

\item{...}{other arguments to be passed to \code{\link{fishStep2D}}}
}
\value{
length-1 list of length-2 named numeric vector indicating rate of change of state variables
}
\description{
Wrapper for 2D version of model for finding stability properties using \code{phaseR} package
}
\seealso{
\code{\link{fishStep2D}}
}

