% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRoot.R
\name{getRoot}
\alias{getRoot}
\title{Get the Root of the Fish Model
Finds the roots of the fish model given starting values (adults, [sun]fish, and juveniles) and parameter (harvest of adults, qE)}
\usage{
getRoot(x, pars, maxiter = 1000, ...)
}
\arguments{
\item{x}{a vector of length 3 with names A0 (water P), F0 (mud P), and J0 (P loading)}

\item{pars}{vector of parameters to pass to model}

\item{maxiter}{maximum number of iterations to use when finding root}

\item{...}{additional arguments to pass to \code{\link{fishStep}}}
}
\value{
numeric vector with roots
}
\description{
Get the Root of the Fish Model
Finds the roots of the fish model given starting values (adults, [sun]fish, and juveniles) and parameter (harvest of adults, qE)
}
\examples{
getRoot(c(A0=1, F0=1, J0=1), pars=c(qE=0.5))
}

