% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRoot.R
\name{getRoot2D}
\alias{getRoot2D}
\title{Get the Root of the Fish Model in 2D}
\usage{
getRoot2D(x, pars, maxiter = 1000, ...)
}
\arguments{
\item{x}{a vector of length 2 with names J0 (juveniles) and F0 (planktivorous fish)}

\item{pars}{vector of parameters to pass to model}

\item{maxiter}{maximum number of iterations to use when finding root}

\item{...}{additional arguments to pass to \code{\link{fishStep2D}}}
}
\value{
numeric vector with roots
}
\description{
Numerically finds the roots of the fish model given starting values (juveniles and [sun]fish) and parameter (harvest of adults, qE)
}
\examples{
getRoot2D(c(F0=1, J0=1), pars=c(qE=0.65))
getRoot2D(c(F0=1, J0=1000), pars=c(qE=0.65))
getRoot2D(c(F0=1, J0=1), pars=c(qE=0.5))
getRoot2D(c(F0=1, J0=1000), pars=c(qE=0.5))
}

