% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRoot.R
\name{getRoot2D_df}
\alias{getRoot2D_df}
\title{Get the Root of the Trophic Cascade Model from a Data Frame}
\usage{
getRoot2D_df(initialValues, maxiter = 1000, ...)
}
\arguments{
\item{initialValues}{a \code{data.frame} with 3 columns named J0 (juvenile bass), F0 (planktivorous fish), and qE (harvest)}

\item{maxiter}{maximum number of iterations}

\item{...}{additional arguments to pass to \code{\link{fishStep2D}}}
}
\value{
a matrix with columns for initial values of J0 and F0, roots, and the parameter qE. Each row of output corresponds to a row of the input (though output reorded to ascending qE). Original columns for state variables (J0 and F0) will be renamed to "init.J0" and "init.F0". The equilibria that would be approached from these initial values, i.e. the 'roots', will take on the column names of "J0" and "F0". Thus, the columns J0 & F0 will not have the same values in the input as in the output (output are equilibrium values), unless the input states were already at equilibrium.
}
\description{
Finds the roots of the trophic cascade model given starting values (juvenile bass, planktivorous fish) and parameter (harvest)
}
\examples{
J0v <- seq(0, 1.5E3, length=2)
F0v <- 20
qEv <- seq(0.15, 1.5, length=10)
iV <- expand.grid(J0=J0v, F0=F0v, qE=qEv)
roots <- getRoot2D_df(iV)
}

