% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasePortrait.R
\name{phasePortrait}
\alias{phasePortrait}
\title{Phase Portrait}
\usage{
phasePortrait(qE, pars, func = fs2D, x.lim = c(0, 2000), y.lim = c(0,
  100), add = FALSE, nFlow = 20, nNull = 300, t.end = 50,
  inits = matrix(c(1, 90, 200, 20, 800, 20, 500, 50, 1000, 80), byrow = TRUE,
  ncol = 2), addLeg = FALSE, legPos = "topright")
}
\arguments{
\item{qE}{numeric scalar for harvest}

\item{pars}{optional named numeric vector for additional parameters to be passed to \code{func}}

\item{func, }{function that returns local derivatives, defaults to \code{\link{fs2D}}, which is just a formatting wrapper for \code{\link{fishStep2D}}; currently only works for default due to use of \code{\link{getRoot2D_df}} in \code{\link{stabClass}}}

\item{x.lim, y.lim}{minimum and maximum values for each of the two states variables to be used when plotting and searching for equilibria}

\item{add}{local, if FALSE (default) makes a new plot}

\item{nFlow}{square root of the number of arrows to draw in the vector field}

\item{nNull}{numeric scalar, number of points used in drawing the nullclines}

\item{t.end}{number of time steps (in model time units) to use when simulating the trajectories}

\item{inits}{matrix of initial values to use for trajectories}

\item{addLeg}{if TRUE, adds a legend}

\item{legPos}{character position of legend; see \code{graphics::legend}}
}
\value{
a data.frame with stability information; see \code{\link{stabClass}}
}
\description{
Plot a phase portrait of a dynamical system that includes the vector field, nullclines, example trajectories, and the identification and classification of equilibria.
}
\details{
This function is largely a wrapper for \code{phaseR::flowField}, \code{phaseR::trajectory}, \code{phaseR::nullclines}, and \code{phaseR::stability}. This function does not call \code{phaseR::stability} directly, and instead uses \code{\link{stabClass}}.
}
\examples{
# shortInits <- eval(formals(phasePortrait)$inits)[1:3,]
# pars <- c(q=8,s=0.7,h=0.15, b=0.001)
qEvals <- c(0.2, 0.5, 1, 1.25)
par(mfrow=c(2,2), mar=c(2,2,1,0.5), mgp=c(1,0.25,0), tcl=-0.15, cex=1, ps=9, cex.axis=0.85)
for(j in 1:4){
	(phasePortrait(
		qE=qEvals[j], pars=NULL,
		nFlow=5, nNull=5,t.end=5,
		addLeg=TRUE
	))
}
}

