% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{stabClass}
\alias{stabClass}
\title{Stability Classification}
\usage{
stabClass(qE, pars, func = fs2D, Jvals = seq(0, 2000, length.out = 2),
  Fvals = seq(10, 20, length.out = 2))
}
\arguments{
\item{qE}{numeric scalar representing harvest rate; reasonable values might be between 0 and 2}

\item{pars}{a named numeric vector of optional additional parameter values to be passed to \code{func}}

\item{func}{function whose stability is to be classified; nominally this is \code{fs2D}, a wrapper for \code{\link{fishStep2D}}; currently this function only works with the default b/c \code{\link{getRoot_df}} hasn't been set up to accept other functions as an argument}

\item{Jvals, Mvals}{numeric vector of J0 (juvenile bass) or F0 (planktivorous fish) values over which to search for equilibria}
}
\value{
a data.frame with a row for each equilibrium point found, and columns for the J0,F0 coordinates of those points, a character describing its stability classification, the trace (tr) and determinant (Delta) of the Jacobian matrix at that point, the 'discriminant' value (tr^2 - 4*delta), and the parameter values supplied through qE and pars.
}
\description{
Find equilibria for a model of a trophic cascade
}
\examples{
stabClass(qE=0.65)
qEvec <- seq(0.15, 1.5, length.out=10)
lout <- lapply(qEvec, stabClass)
do.call(rbind, lout)
}

